text_analyzer/pyproject.toml
[tool.poetry]
name = "text-analyzer"
version = "0.1.0"
description = ""
authors = ["Vladislav Sokolov <coko.2002@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
nltk = "^3.8.1"
pymorphy3 = "^1.3.1"
numpy = "^1.26.2"
pyyaml = "^6.0.1"
scikit-learn = "1.0.2"
requests = "^2.31.0"
pika = "^1.3.2"
python-docx = "^1.1.0"
docker = "^7.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
